+++ app/brain/classifier.py
@@ -55,6 +55,7 @@ Pesan user:
             "reservation",
             "view_reservation",
             "update_reservation",
+            "cancel_reservation",
             "menu",
             "promo",
             "faq",